Moved createMatchMedia.js to test/helpers
